persistencia entidad vehiculo

+++ src/main/java/com/progra3/proyecto/entity/Vehiculo.java
@@ -10,19 +10,17 @@ public class Vehiculo {
 	 
 		@Id
 		@GeneratedValue(strategy = GenerationType.IDENTITY)	
-		private Integer id;
+		private Long id;
 		private String tipo;		
 		private String nombre;		
 		private String color;
 
-		public Integer getId() {
+		public Long getId() {
 			return id;
 		}
-
-		public void setId(Integer id) {
+		public void setId(Long id) {
 			this.id = id;
 		}
-
 		public String getTipo() {
 			return tipo;
 		}
